+++ servers/cromwell/jobs/controllers/jobs_controller.py
@@ -213,11 +213,13 @@ def format_task(task_name, task_metadata):
 
 def format_task_failure(task_name, metadata):
     return FailureMessage(task_name=remove_workflow_name(task_name),
-                          failure=metadata['failures'][0].get('message'),
+                          failure=get_deepest_message(
+                              metadata.get('failures')),
                           timestamp=_parse_datetime(metadata.get('end')),
                           stdout=metadata.get('stdout'),
                           stderr=metadata.get('stderr'),
-                          call_root=metadata.get('callRoot'))
+                          call_root=metadata.get('callRoot'),
+                          job_id=metadata.get('subWorkflowId'))
 
 
 def format_workflow_failure(failures):
@@ -441,6 +443,14 @@ def _get_execution_events(events):
             for event in events
         ]
     return execution_events
+
+
+def get_deepest_message(metadata_list):
+    if 'causedBy' in metadata_list[0] and len(
+            metadata_list[0].get('causedBy')):
+        return get_deepest_message(metadata_list[0].get('causedBy'))
+    else:
+        return metadata_list[0].get('message')
 
 
 def _parse_datetime(date_string):